chore: enum 파일 분리
